requirements.txt
appdirs
hidapi
pyobjc
